Refactor frontend app shell for issue #45
