+++ ProjectPortfolio2026/projectportfolio2026.client/src/app/navigation.ts
@@ -1,0 +1,6 @@
+export interface NavigationOptions {
+    replace?: boolean;
+    preserveScroll?: boolean;
+}
+
+export type NavigateFn = (path: string, options?: NavigationOptions) => void;
